fixed makefiles and warnings
Former-commit-id: eff24004d4f83fd4f1d89b818c2c031d7b2ed4e9

+++ lib/ncurses/include/Graphical/Window.hpp
@@ -43,7 +43,7 @@ class Window : public IWindow {
         void  initColors(void);
         short getColor(Color);
         int   getColorPair(Color, Color);
-        void  drawBufferPixel(int, int, Color);
+        void  drawBufferPixel(size_t, size_t, Color);
         std::vector<std::vector<Color>> getBuffer(void);
 
     protected:
